requirements.txt
torch>=2.4.0
deepspeed>=0.16.0
hydra-core>=1.3.2
omegaconf>=2.3.0
tensorboardX>=2.6.2.2
numpy>=1.26.3
transformers>=4.46.3
soundfile>=0.12.1
librosa>=0.10.2.post1
